refactor: give the chunks schema to the memory that holds it
The shape of the `chunks` dataset — its schema, the Chunk → RecordBatch conversion, the
embedding model it's pinned to and the vector width — lived in the `index` command, so
`memory::hub::check_compat` had to import DIM and MODEL back out of a command to decide
whether a memory is readable. That was the crate's only domain → command dependency.

The schema, build_batch, MODEL and DIM move to memory::dataset, next to the table name
and the open/scan helpers that already own the dataset; the column-order tripwire test
moves with them. embed_batched and its batch size go to inference, whose Embedder trait
is all they touch.

No logic changes: the moved bodies are unchanged, and every caller (index, scrub, curate,
push) now names the module that owns what it asks for.
