+++ src/memory/dataset.rs
@@ -1,13 +1,16 @@
-//! Shared memory helpers: the local memory location, opening a dataset, plain scans, and building the
-//! FTS/IVF indexes. funes's home is `$FUNES_HOME`/`~/.funes` — it holds the incremental state and
-//! the local memory at `…/memory` (the `chunks` Lance dataset).
+//! Shared memory helpers: the local memory location, the `chunks` table's schema and rows, opening a
+//! dataset, plain scans, and building the FTS/IVF indexes. funes's home is `$FUNES_HOME`/`~/.funes` —
+//! it holds the incremental state and the local memory at `…/memory` (the `chunks` Lance dataset).
 
+use crate::chunk;
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::Arc;
 
 use anyhow::{Context, Result};
-use arrow_array::RecordBatch;
+use arrow_array::types::Float32Type;
+use arrow_array::{FixedSizeListArray, Int64Array, RecordBatch, StringArray};
+use arrow_schema::{DataType, Field, Schema};
 use futures::TryStreamExt;
 use lance::dataset::builder::DatasetBuilder;
 use lance::dataset::Dataset;
@@ -22,6 +25,12 @@ use lance_linalg::distance::MetricType;
 
 /// The table (Lance dataset) name within a memory.
 pub const TABLE: &str = "chunks";
+
+/// The embedding model a memory's vectors are built with, and their width. Pinned in the schema
+/// metadata and enforced on open ([`super::hub::Memory::open`]): a memory built with another model
+/// can't be queried with funes's embeddings.
+pub const MODEL: &str = "BAAI/bge-small-en-v1.5";
+pub const DIM: i32 = 384;
 
 /// funes's home directory: `$FUNES_HOME`, else `~/.funes`. Holds the incremental state and the
 /// local memory.
@@ -154,4 +163,107 @@ fn ivf_pq_params(ds: &Dataset) -> Option<VectorIndexParams> {
         IvfBuildParams::default(),
         pq,
     ))
+}
+
+/// The table schema (column order is load-bearing for Lance).
+pub(crate) fn schema() -> Arc<Schema> {
+    let utf8 = |name: &str| Field::new(name, DataType::Utf8, true);
+    let i64f = |name: &str| Field::new(name, DataType::Int64, true);
+    Arc::new(Schema::new_with_metadata(
+        vec![
+            utf8("id"),
+            utf8("text"),
+            utf8("session_id"),
+            utf8("workdir"),
+            utf8("turn_uuid"),
+            utf8("parent_uuid"),
+            i64f("seq"),
+            utf8("ts"),
+            utf8("role"),
+            utf8("block_type"),
+            utf8("tool_name"),
+            utf8("source_path"),
+            i64f("block_idx"),
+            i64f("split_idx"),
+            Field::new(
+                "vector",
+                DataType::FixedSizeList(Arc::new(Field::new("item", DataType::Float32, true)), DIM),
+                true,
+            ),
+            // After `vector`: `add_columns` appends a migrated column at the end, so a
+            // freshly-built memory must match that order (the tripwire test pins it). `harness`
+            // came first, then `repo` — each appended in turn.
+            utf8("harness"),
+            utf8("repo"),
+        ],
+        HashMap::from([("embedding_model".to_string(), MODEL.to_string())]),
+    ))
+}
+
+pub(crate) fn build_batch(chunks: &[chunk::Chunk], vectors: &[Vec<f32>]) -> Result<RecordBatch> {
+    let s = |f: &dyn Fn(&chunk::Chunk) -> Option<String>| -> StringArray { chunks.iter().map(f).collect() };
+    let i = |f: &dyn Fn(&chunk::Chunk) -> i64| -> Int64Array { chunks.iter().map(|c| Some(f(c))).collect() };
+    let vector = FixedSizeListArray::from_iter_primitive::<Float32Type, _, _>(
+        vectors
+            .iter()
+            .map(|v| Some(v.iter().map(|&x| Some(x)).collect::<Vec<_>>())),
+        DIM,
+    );
+    Ok(RecordBatch::try_new(
+        schema(),
+        vec![
+            Arc::new(s(&|c| Some(c.id.clone()))),
+            Arc::new(s(&|c| Some(c.text.clone()))),
+            Arc::new(s(&|c| Some(c.session_id.clone()))),
+            Arc::new(s(&|c| Some(c.workdir.clone()))),
+            Arc::new(s(&|c| Some(c.turn_uuid.clone()))),
+            Arc::new(s(&|c| c.parent_uuid.clone())),
+            Arc::new(i(&|c| c.seq)),
+            Arc::new(s(&|c| Some(c.ts.clone()))),
+            Arc::new(s(&|c| Some(c.role.clone()))),
+            Arc::new(s(&|c| Some(c.block_type.clone()))),
+            Arc::new(s(&|c| c.tool_name.clone())),
+            Arc::new(s(&|c| Some(c.source_path.clone()))),
+            Arc::new(i(&|c| c.block_idx)),
+            Arc::new(i(&|c| c.split_idx)),
+            Arc::new(vector),
+            Arc::new(s(&|c| Some(c.harness.clone()))),
+            Arc::new(s(&|c| Some(c.repo.clone()))),
+        ],
+    )?)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn schema_column_order_is_load_bearing() {
+        // Column order must match build_batch's array order exactly, or Lance writes the
+        // wrong column. Pin it so a reorder can't slip through.
+        let s = schema();
+        let names: Vec<&str> = s.fields().iter().map(|f| f.name().as_str()).collect();
+        assert_eq!(
+            names,
+            vec![
+                "id",
+                "text",
+                "session_id",
+                "workdir",
+                "turn_uuid",
+                "parent_uuid",
+                "seq",
+                "ts",
+                "role",
+                "block_type",
+                "tool_name",
+                "source_path",
+                "block_idx",
+                "split_idx",
+                "vector",
+                "harness",
+                "repo",
+            ]
+        );
+    }
 }
