+++ tests/push_round_trip.rs
@@ -226,7 +226,7 @@ async fn push_round_trip_create_append_recall() {
     );
     assert_eq!(
         remote_model.as_deref(),
-        Some(funes::commands::index::MODEL),
+        Some(funes::memory::dataset::MODEL),
         "the model id should travel with the memory via push"
     );
     assert_eq!(
